// src/pages/sign-up/sign-up.d.ts
export interface ISignUpMutateValues {
  name: string;
  surname: string;
  country: string;
  city: string;
  address: string;
  liveAddress: string;
  company: string;
  phone: string;
  email: string;
  password: string;
  catUser: string;
};

export interface IFormValues {
  name: string;
  surname: string;
  country: string;
  city: string;
  address: string;
  liveAddress: string;
  company: string;
  phone: string;
  email: string;
  password: string;
}